Implemented a physical memory manager

+++ kernel/src/memory/memory.h
@@ -1,9 +1,10 @@
 #pragma once
 #include <efiMemory.h>
 #include <stdint.h>
+#include <stddef.h>
 
 uint64_t GetMemorySize(EFI_MEMORY_DESCRIPTOR* mMap, uint64_t mMapEntries, uint64_t mDescriptorSize);
 
-void* memcpy(void* dst, const void* src, uint16_t num);
-void* memset(void* ptr, int value, uint16_t num);
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
+void* memcpy(void* dst, const void* src, size_t num);
+void* memset(void* ptr, int value, size_t num);
+int memcmp(const void* ptr1, const void* ptr2, size_t num);
